feat: TurnDao 인터페이스 및 구현체 수정
